requirements.txt
colorama
requests
fade
pyyaml